Fix warning with gcc 5.4

+++ media/libstagefright/codecs/mp3dec/SoftMP3.cpp
@@ -265,7 +265,7 @@ void SoftMP3::onQueueFilled(OMX_U32 /* portIndex */) {
         mConfig->outputFrameSize = kOutputBufferSize / sizeof(int16_t);
         if ((int32)outHeader->nAllocLen < mConfig->outputFrameSize) {
             ALOGE("input buffer too small: got %lu, expected %u",
-                outHeader->nAllocLen, mConfig->outputFrameSize);
+                (unsigned long)outHeader->nAllocLen, mConfig->outputFrameSize);
             android_errorWriteLog(0x534e4554, "27793371");
             notify(OMX_EventError, OMX_ErrorUndefined, OUTPUT_BUFFER_TOO_SMALL, NULL);
             mSignalledError = true;
